Making the demo application a bit more real.
In order to seem more real we add a couple of common bundles to the
demo application.  The first bundle is the version bundle -- with
this bundle the application will listen to a `/version` request on
the admin port and return with a JSON object that describes all of
the dependencies compiled into the jar as well as their versions.
The second bundle is the datadog metrics bundle -- with this bundle
the application can include configuration in its config file to
have metrics periodically sent to datadog.