Prepare for ECS deployment.
In order to prepare for deploying in ECS we do a few things to the
demo project:

1. Update the pom to specify the docker artifact name.

2. Read the configuration file from the classpath instead of from
disk.  A new config file has been checked into src/main/resources
for this purpose.  Also the newly checked in config file has been
made a template.  This gives it the ability to read values from
the system environment and use those values in configuration
properties.

3. Switch over to the ECS version of the Datadog metrics bundle.
This new version of the bundle is ECS aware and will pull various
properties from the ECS environment and send them as tags on
each metric sent to Datadog.

4. JSON logging has been enabled in the templated config file,
but is controllable via the LOG_FORMAT environment variable.

5. `/ping` is mapped to the PingServlet in the same way it works
on the admin port.  This is necessary because in ECS the load
balancer only supports a single port.